update code extract schemas from model
